Optimistic-er commit
