+++ services/app/src/actions/Audio.tsx
@@ -1,16 +1,34 @@
-import {NodeSet} from '../components/base/Audio';
-import {AudioState} from '../reducers/StateTypes';
-import {AudioSetAction} from './ActionTypes';
+import * as Redux from 'redux';
+import {AudioNode} from '../audio/AudioNode';
+import {ThemeManager} from '../audio/ThemeManager';
+import {MUSIC_DEFINITIONS} from '../Constants';
+import {getAudioContext} from '../Globals';
+import {AudioDataState, AudioState} from '../reducers/StateTypes';
+import {AudioDataSetAction, AudioSetAction} from './ActionTypes';
+
+export function getAllMusicFiles(): string[] {
+  return Object.keys(MUSIC_DEFINITIONS).reduce((list: string[], musicClass: string) => {
+    return list.concat(Object.keys(MUSIC_DEFINITIONS[musicClass]).reduce((acc: string[], musicWeight: string) => {
+      const weight = MUSIC_DEFINITIONS[musicClass][musicWeight];
+      for (const instrument of weight.instruments) {
+        for (let v = 1; v <= weight.variants; v++) {
+          acc.push(`${musicClass}/${musicWeight}/${instrument}${v}`);
+        }
+      }
+      return acc;
+    }, []));
+  }, []);
+}
 
 // can't use Fetch for local files since audio files might come from file://, must use this instead
 // TODO: Switch to using promises
-export function loadAudioLocalFile(context: AudioContext, url: string, callback: (err: Error|null, buffer: NodeSet|null) => void) {
+export function loadAudioLocalFile(context: AudioContext, url: string, callback: (err: Error|null, buffer: AudioNode|null) => void) {
   const request = new XMLHttpRequest();
   request.open('GET', url, true);
   request.responseType = 'arraybuffer';
   request.onload = () => {
     context.decodeAudioData(request.response, (buffer: AudioBuffer) => {
-      return callback(null, new NodeSet(context, buffer));
+      return callback(null, new AudioNode(context, buffer));
     }, (err: Error) => {
       return callback(err, null);
     });
@@ -19,6 +37,39 @@ export function loadAudioLocalFile(context: AudioContext, url: string, callback:
     return callback(Error('Network error'), null);
   };
   request.send();
+}
+
+function audioDataSet(data: Partial<AudioDataState>): AudioDataSetAction {
+  return {type: 'AUDIO_DATA_SET', data};
+}
+
+export function loadAudioFiles() {
+  return (dispatch: Redux.Dispatch<any>): any => {
+    const ac = getAudioContext();
+    if (!ac) {
+      return;
+    }
+
+    console.log('Starting audio load');
+    dispatch(audioSet({loaded: 'LOADING'}));
+    const musicFiles = getAllMusicFiles();
+    // TODO: eachLimit
+    const audioNodes: {[key: string]: AudioNode} = {};
+    for (const file of musicFiles) {
+      loadAudioLocalFile(ac, 'audio/' + file + '.mp3', (err: Error|null, ns: AudioNode) => {
+        if (err) {
+          dispatch(audioSet({loaded: 'ERROR'}));
+          console.error('Error loading audio file: ' + file);
+        } else {
+          audioNodes[file] = ns;
+        }
+      });
+    }
+    dispatch(audioSet({loaded: 'LOADED'}));
+    const themeManager = new ThemeManager(audioNodes, Math.random);
+    dispatch(audioDataSet({audioNodes, themeManager}));
+    console.log('Audio loaded');
+  };
 }
 
 export function audioSet(delta: Partial<AudioState>): AudioSetAction {
